+++ clean_database.py
@@ -1,0 +1,5 @@
+import database
+db = database.DataBase()
+
+
+db.insertTestData()